requirements.txt
streamlit==1.31.0
numpy>=1.26.0
opencv-python-headless==4.8.1.78
Pillow==10.0.0
pandas>=2.1.0
altair>=5.0.0